Updates database.js to expose query functions and adds test fixture support.
Since database.js is the file that gets exported, the functions for querying
and transactions need to be exposed. This commit adds .query and
.performTransaction functions with a check to see if the database has been
connected in both. Supporting tests have been written.

The connect function was modified to allow for a new option "useFixture" which
adds test fixture support to table models. This option should be used for test
environments. Supporting tests have been written.