Updates test scripts to support integration tests and modifies db interface.
Created seperated tests scripts for unit and integration tests in package.json
Modified the connect function to return the database object instead of true
and also modifies the error handler to let the error be thrown up the chain.
Renamed database to database.singleton to more accurately describe the code
contained within.