Adds code to support size definition when selecting a column type.
Postgres allows for the user to define the size of certain types. Code
was added to allow size to be defined similiarly as the sql commands would be.
The size gets added as an option to the column definition. This value will be
used later for other features.
Supporting tests were updated.