Integration tests were written for the test fixture functions.
During this refactoring the following bugs were found and corrected.
* bigint returns properly as a string following the pg module return style.
Code generating the bigint for tests was updated and the unit and integration
tests were corrected.
* Date types were not being generated correctly. The generator code was
updated to return a date object with the time zeroed out.
* The delete functions in the test.fixtures file did not actually work
as written. The sql statements were updated to how the pg module likes them.
The code that existed was attempting to be "smart" about how to delete
dynmically however it did not work.